backend/fastapi_app/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
supabase==1.0.0
